new_www/pyproject.toml
[tool.poetry]
name = "new-www"
version = "0.1.0"
description = ""
authors = ["Mariusz Drabik <drabik.mariusz@gmail.com>"]
readme = "README.md"
packages = [{include = "new_www"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.23"
pydantic-settings = "^2.0.3"
pycodestyle = "^2.11.1"
psycopg2 = "^2.9.9"
alembic = "^1.12.1"
bcrypt = "^4.0.1"
email-validator = "^2.1.0.post1"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
